GreatestCommonDivisorCalculator
In this project, we will create a Java program that finds the EBOB (Greatest Common Divisor) of two numbers entered by the user. EBOB is the greatest value between the common divisors of two numbers. In this project, we will ask the user to enter two numbers and calculate the EBOB for these numbers. Euclidean algorithm will be used to calculate EBOB.